# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Farmers Market Platform API",
    "version": "1.0.0",
    "description": "Divine Agricultural Platform - Comprehensive REST API\n\nThis API provides endpoints for:\n- 🚜 Farm Management (CRUD, search, verification)\n- 🌾 Product Catalog (inventory, categories, search)\n- 📦 Order Processing (cart, checkout, fulfillment)\n- 👤 User Management (authentication, profiles, roles)\n- 🌍 Marketplace (discovery, featured items, recommendations)\n- 📊 Analytics (farmer dashboard, sales reports)\n- 🔔 Notifications (real-time updates, alerts)\n\nAll endpoints return standardized ServiceResponse<T> format with:\n- success: boolean\n- data?: T\n- error?: ServiceError\n- meta?: ResponseMetadata (including agricultural consciousness)",
    "contact": {
      "name": "API Support",
      "email": "support@farmersmarket.com",
      "url": "https://farmersmarket.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["ProductController"],
        "summary": "handleListProducts",
        "description": "List all products with pagination and filtering",
        "operationId": "get__api_products",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "farmId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of products"
          }
        }
      },
      "post": {
        "tags": ["ProductController"],
        "summary": "handleCreateProduct",
        "description": "Create a new product (requires FARMER role and farm ownership)",
        "operationId": "post__api_products",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        }
      }
    },
    "/api/products/search": {
      "get": {
        "tags": ["ProductController"],
        "summary": "handleSearchProducts",
        "description": "Search products by keyword",
        "operationId": "get__api_products_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from NextAuth"
      }
    },
...
  }
}